refactor: added penalites schema and interface
