+++ github/copilot-instructions.md
@@ -1,0 +1,1 @@
+Concrete JSON types for TypeScript usage (no `any`, `unknown`, or explicit `undefined`)
