+++ packages/data/schema/system/src/model/SSS/file_and_storage/storage_config.schema.ts
@@ -1,5 +1,5 @@
 import mongoose, { Schema } from 'mongoose';
-import { IStorageConfig } from './file_and_storage.interfaces';
+import { IStorageConfig } from './interface';
 import { randomUUID } from 'crypto';
 
 const StorageConfigSchema = new Schema<IStorageConfig>(
