+++ packages/data/schema/system/src/model/SSS/inter_agency/interagency.schema.ts
@@ -1,0 +1,31 @@
+// @ts-ignore
+import { Schema, model } from 'mongoose';
+import { IInteragencyConfig } from './interface';
+
+const InteragencySchema = new Schema<IInteragencyConfig>(
+    {
+        id: { type: Number, required: false, sparse: true, index: true },
+        uuid: { type: String, required: true, unique: true, index: true },
+        agencyCode: { type: String, trim: true, index: true },
+        agencyName: { type: String, trim: true },
+        description: { type: String },
+        integrationSettings: { type: Schema.Types.Mixed, default: {} },
+        accessControl: { type: Schema.Types.Mixed, default: {} },
+        workflowParticipation: { type: Schema.Types.Mixed, default: {} },
+        logging: { type: Schema.Types.Mixed, default: {} },
+        status: { type: String, default: 'inactive' },
+        lastSynced: { type: Date },
+        createdBy: { type: Schema.Types.Mixed },
+    },
+    {
+        timestamps: { createdAt: 'createDate', updatedAt: 'updateDate' },
+        collection: 'interagencyConfig',
+    }
+);
+
+// unique index on agencyCode where present (sparse) to avoid collisions for documents without the field
+InteragencySchema.index({ agencyCode: 1 }, { unique: true, sparse: true });
+
+export const InteragencyModel = model<IInteragencyConfig>('InteragencyConfig', InteragencySchema);
+
+export default InteragencyModel;
