+++ packages/data/schema/system/src/model/SSS/penalties/penalty_computations_config.ts
@@ -1,0 +1,35 @@
+//@ts-ignore
+import { Schema, model } from 'mongoose';
+import { IPenaltyComputationConfig } from './interface';
+
+const PenaltyComputationConfigSchema = new Schema<IPenaltyComputationConfig>(
+  {
+    uuid: { type: String, required: true, unique: true, index: true },
+    env: { type: String },
+    rounding: { type: Schema.Types.Mixed },
+    validation: { type: Schema.Types.Mixed },
+    defaultStatus: { type: Schema.Types.Mixed },
+    lockThresholdDays: { type: Number },
+    allowGracePeriod: { type: Boolean, default: false },
+    gracePeriodDays: { type: Number },
+    fallbackPenaltyRate: { type: Number },
+    auditTrailEnabled: { type: Boolean, default: false },
+    logComputationDetails: { type: Boolean, default: false },
+
+    // Relations (store ObjectId references and optional embedded snapshot)
+    feesConfigId: { type: Schema.Types.ObjectId, ref: 'FeesConfig' },
+    penaltyConfigId: { type: Schema.Types.ObjectId, ref: 'PenaltyConfig' },
+    feesConfig: { type: Schema.Types.Mixed },
+    penaltyConfig: { type: Schema.Types.Mixed },
+
+    createdBy: { type: Schema.Types.ObjectId, ref: 'User' },
+  },
+  { timestamps: { createdAt: 'createdAt', updatedAt: 'updatedAt' } }
+);
+
+export const PenaltyComputationConfigModel = model<IPenaltyComputationConfig>(
+  'PenaltyComputationConfig',
+  PenaltyComputationConfigSchema
+);
+
+export default PenaltyComputationConfigModel;
