+++ packages/data/schema/system/src/model/SSS/penalties/penalty_config.schema.ts
@@ -1,0 +1,18 @@
+//@ts-ignore
+import { Schema, model } from 'mongoose';
+import { IPenaltyConfig } from './interface';
+
+const PenaltyConfigSchema = new Schema<IPenaltyConfig>(
+  {
+    uuid: { type: String, required: true, unique: true, index: true },
+    licenseType: { type: String },
+    tiers: { type: Schema.Types.Mixed },
+    lockThreshold: { type: Number },
+    createdBy: { type: Schema.Types.ObjectId, ref: 'User' },
+  },
+  { timestamps: { createdAt: 'createdAt', updatedAt: 'updatedAt' } }
+);
+
+export const PenaltyConfigModel = model<IPenaltyConfig>('PenaltyConfig', PenaltyConfigSchema);
+
+export default PenaltyConfigModel;
